Updated ui.sh, launcher.sh, and debloat.sh (Linux/macOS) 07/13/2022
- Added Alexa Cards to Debloat.txt
- Switched to case for interface functions (ui, launcher, & debloat)
- Simplified device identifier
- Unknown devices are now "Unsupported/Unknown Devices"
- Added `-r` argument to appinstaller
- Fixed Disable OTA function not finishing
- ui.sh now passes shellcheck without warnings

+++ Fire-Tools/ui.sh
@@ -3,21 +3,23 @@
 # Check for ADB
 command -v adb >/dev/null 2>&1 || { echo >&2 "This application requires ADB to be installed, Exiting..."; exit 1; }
 
-# Get Device Name from Amazon Docs Using Model Number, if Device Not Connected Set Model to Null. Cache Website for Subsequent Searches
-model=$(adb shell getprop ro.product.model) || model="Null"
-[ -e identifying-tablet-devices.html ] ||
-curl -o ./identifying-tablet-devices.html "https://developer.amazon.com/docs/fire-tablets/ft-identifying-tablet-devices.html"
-device=$(grep -B 2 "$model" < identifying-tablet-devices.html | grep "Fire" | awk '$0=$2' FS=">" RS="<")
-[ "$device" = "" ] && device="Unknown or Unsupported Device"
+# If Device is Connected Identify Using Amazon Docs Page
+device="Device not Detected"
+if (adb shell echo "Device Connected"); then
+    model=$(adb shell getprop ro.product.model)
+    [ -e identifying-tablet-devices.html ] || curl -o ./identifying-tablet-devices.html "https://developer.amazon.com/docs/fire-tablets/ft-identifying-tablet-devices.html"
+    device=$(grep -B 2 "$model" < identifying-tablet-devices.html | grep "Fire" | awk '$0=$2' FS=">" RS="<")
+    [ -z "$device" ] && device="Unsupported/Unknown Device"
+fi
 
-# Change Application Installation Method Based on the Type of File
+# Change Application Installation Method Based on Filetype
 appinstaller () {
     case "$1" in
     *.apk)
-        adb install -g "$1";;
+        adb install -r -g "$1";;
     *.apkm)
         unzip "$1" -d ./Split
-        adb install-multiple -g ./Split/*.apk
+        adb install-multiple -r -g ./Split/*.apk
         rm -rf ./Split;;
     esac
 }
@@ -35,59 +37,54 @@ tool=$(zenity --list \
     "Batch Installer" "Install all .apk(m) files in the Batch folder" \
     "Update" "Grab the latest Fire-Tools scripts")
 
-# Debloat Menu
-[ "$tool" = "Debloat" ] && exec ./debloat.sh
+# Tool Functions
+case "$tool" in
+    "Debloat")
+        exec ./debloat.sh;;
 
-# Install Google Services 
-if [ "$tool" = "Google Services" ]; then
-    for gapps in ./Gapps/Google*.apk*
-    do
-        appinstaller "$gapps"
-    done
-    echo "Successfully Installed Google Apps"
-fi
+    "Google Services")
+        for gapps in ./Gapps/Google*.apk*
+        do
+            appinstaller "$gapps"
+        done
+        echo "Successfully Installed Google Apps, Re-run Tool if Play Store Crashes";;
 
-# Custom Launcher Menu
-[ "$tool" = "Change Launcher" ] && exec ./launcher.sh
+    "Change Launcher")
+        exec ./launcher.sh;;
 
-# Disable OTA Updates
-if [ "$tool" = "Disable OTA" ]; then
-    export ota="com.amazon.device.software.ota com.amazon.device.software.ota.override com.amazon.kindle.otter.oobe.forced.ota"
-    for package in ${ota}
-    do
-        adb shell pm disable-user -k "$package" ||
-        echo "Failed to Disable OTA Updates"; exec ./ui.sh
-    done
-    echo "Successfully Disabled OTA Updates"
-fi 
+    "Disable OTA")
+        export ota="com.amazon.device.software.ota com.amazon.device.software.ota.override com.amazon.kindle.otter.oobe.forced.ota"
+        for package in ${ota}
+        do
+            adb shell pm disable-user -k "$package" ||
+            { echo "Failed to Disable OTA Updates"; exec ./ui.sh; }
+        done
+        echo "Successfully Disabled OTA Updates";;
 
-# Select Package from List & Extract Package's APK file(s)
-[ "$tool" = "Apk Extractor" ] &&
-    adb shell pm list packages | cut -f 2 -d ":" > packagelist &&
-    package=$(zenity --list --width=500 --height=400 --column=Packages < packagelist) &&
-    adb shell pm path "$package" | cut -f 2 -d ":" > Packages.txt &&
-    xargs -L1 adb pull < Packages.txt &&
-    echo "Finished Extracting App(s)"
+    "Apk Extractor")
+        adb shell pm list packages | cut -f 2 -d ":" > packagelist
+        package=$(zenity --list --width=500 --height=400 --column=Packages < packagelist) &&
+        adb shell pm path "$package" | cut -f 2 -d ":" > Packages.txt &&
+        xargs -L1 adb pull < Packages.txt &&
+        echo "Finished Extracting App(s)";;
 
-# Batch Install
-if [ "$tool" = "Batch Installer" ]; then
-    for apps in ./Batch/*.apkm
-    do
-        appinstaller "$apps"
-    done
-    echo "Finished Installing App(s)"
+    "Batch Installer")
+        for apps in ./Batch/*.apkm
+        do
+            appinstaller "$apps"
+        done
+        echo "Finished Installing App(s)";;
 
-# Updater Tool
-elif [ "$tool" = "Update" ]; then
-    echo "Latest Changelog:" && curl -sSL https://github.com/mrhaydendp/Fire-Tools/raw/main/Changelog.md
-    export modules="Debloat.txt ui.sh debloat.sh launcher.sh"
-    for module in ${modules}
-    do
-        curl -sSL https://github.com/mrhaydendp/Fire-Tools/raw/main/Fire-Tools/"$module" > "$module"
-    done
-    rm ./identifying-tablet-devices.html
-    echo "Successfully Updated, Restarting Now..."
-fi
+    "Update")
+        echo "Latest Changelog:" && curl -sSL https://github.com/mrhaydendp/Fire-Tools/raw/main/Changelog.md
+        export modules="Debloat.txt ui.sh debloat.sh launcher.sh"
+        for module in ${modules}
+        do
+            curl -sSL https://github.com/mrhaydendp/Fire-Tools/raw/main/Fire-Tools/"$module" > "$module"
+        done
+        rm ./identifying-tablet-devices.html
+        echo "Successfully Updated, Reloading Application...";;
+esac
 
 # Exit to Menu When Tool Finishes
-[ "$tool" != "" ] && exec ./ui.sh
+[ -z "$tool" ] || exec ./ui.sh
